high res timer, debugging/logging with JUCE logger, WIN32

+++ src/Widgets/ShapeWidget.h
@@ -61,5 +61,5 @@ class ShapeWidget : public Widget
     double _blueTextColor;
     double _alphaTextColor;
 
-	 double _outlineThickness;
+	 int _outlineThickness;
 };
